Move MigrationTool interfaces into its subdirectory
