+++ examples/utils/migration-tools/src/interfaces/index.ts
@@ -9,12 +9,6 @@ export {
 	IVersionedModel,
 } from "./migratableModel.js";
 export {
-	IAcceptedMigrationDetails,
-	IMigrationTool,
-	IMigrationToolEvents,
-	MigrationState,
-} from "./migrationTool.js";
-export {
 	DataTransformationCallback,
 	IMigrator,
 	IMigratorEvents,
