+++ examples/utils/migration-tools/src/migrationTool/migrationTool.ts
@@ -32,7 +32,7 @@ import type {
 	IMigrationTool,
 	IMigrationToolEvents,
 	MigrationState,
-} from "../interfaces/index.js";
+} from "./interfaces.js";
 
 const consensusRegisterCollectionId = "consensus-register-collection";
 const pactMapId = "pact-map";
